+++ seanime-web/src/app/(main)/_features/video-core/video-core-screenshot.ts
@@ -32,7 +32,7 @@ export function useVideoCoreScreenshot() {
             libassRenderer.resize(true, canvas.width, canvas.height)
             screenshotTimeout.current = setTimeout(() => {
                 ctx.drawImage(libassRenderer._canvas, 0, 0, canvas.width, canvas.height)
-                libassRenderer.resize(true, 0, 0, 0)
+                libassRenderer.resize(true, 0, 0)
                 resolve()
             }, 300)
         })