+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file.
 - ⚡️ Home Screen: Added 'Repeating' filter for 'My lists'
 - ⚡️ Windows: Use user-default explorer application
 - ⚡️ Denshi: Right-click context menu for inputs
+- ⚡️ VideoCore: Increased subtitle parsing speed
 - 🦺 Denshi: Fixed blank screen after server reconnection
 - 🦺 Transcoder: Fixed 10bit issues with NVENC (?)
 - 🦺 Continuity: Fixed potential progress being ignored